! Fix component and rename component
