+++ src/components/kid-metered-paywall/readme.md
@@ -1,0 +1,17 @@
+# kompas-paywall-meter
+
+
+
+<!-- Auto Generated Below -->
+
+
+## Properties
+
+| Property           | Attribute           | Description                                                                        | Type     | Default |
+| ------------------ | ------------------- | ---------------------------------------------------------------------------------- | -------- | ------- |
+| `countdownArticle` | `countdown-article` | prop countdownArticle untuk menghandle sudah berapa artikel gratis yang user baca. | `number` | `0`     |
+
+
+----------------------------------------------
+
+*Terbikin oleh tim front-end kompas.id*
