--- src/components/kompas-paywall-meter/readme.md
@@ -1,19 +1,0 @@
-# kompas-paywall-meter
-
-
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-| Property           | Attribute           | Description                                                                        | Type                    | Default     |
-| ------------------ | ------------------- | ---------------------------------------------------------------------------------- | ----------------------- | ----------- |
-| `countdownArticle` | `countdown-article` | prop countdownArticle untuk menghandle sudah berapa artikel gratis yang user baca. | `number`                | `0`         |
-| `isLogin`          | `is-login`          |                                                                                    | `boolean`               | `false`     |
-| `type`             | `type`              |                                                                                    | `"epaper" \| "reguler"` | `'reguler'` |
-
-
-----------------------------------------------
-
-*Terbikin oleh tim front-end kompas.id*
