+++ src/index.html
@@ -84,7 +84,7 @@
       exercitationem, reiciendis vitae!</p>
     <!-- <button class="btn btn-primary mb-5">tes bootstrap class</button> -->
   </div>
-  <kompas-paywall-meter></kompas-paywall-meter>
+  <kid-metered-paywall countdown-article=6></kid-metered-paywall>
   <div>
     <kompas-footer />
     <!-- <kompas-footer wording-message="wadidaw"/> -->
